requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
nulltype==2.3.1
ory-hydra-client==1.9.2
protobuf==3.14.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
six==1.15.0
urllib3==1.26.3
Werkzeug==1.0.1
